fix(prompts): drop support for pydantic v1 (#6181)
